Use JSON as default file session store
It's always better to have a human-readable JSON file than
an unreadable DAT one containing bytes.

You can see what the session file contains, what it is all
about.

Also, it prevents the file from getting corrupted. Happened
to me a few times in the past using FileSessionStore, had
to delete the session file and reconnect/auth.

And it's just way better in general for storing data. For
example, I currently store a lot of sessions in a SQL
database, and doing it with bytes would have been a pain.

JSON session file is just better in general.

NOTE: Reason for migrating to .NET4.6.2 is to be able to
reference .NETStandard2.0 (needed for System.Text.Json).

Co-authored-by: Andres G. Aragoneses <knocte@gmail.com>